requirements.txt
pandas>=2.2,<3.0
numpy>=1.26,<3.0
streamlit>=1.56,<2.0
plotly>=6.0,<7.0
yfinance>=0.2.54,<1.0
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
kaleido>=0.2.1,<1.0
pytest>=8.0,<9.0
